Fixed (longstanding?) bug in cb_truncate().
The code generated when a program is loaded (with vm_load()) more than once in
a single VM instance would be completely wrong due to cb_truncate() not clearing
the size.  This also caused listing.py to give meaningless output because the
ELF object generated with bfi -c -t would contain garbage.