Test for zero before executing optimized loop code.
Code like this: [-<+>] used to be compiled as:

    movzbq (%rax), %rcx
    add    %cl, -1(%rax)
    movb   $0x0, (%rax)

This code is executed unconditionally, even if the value under the tape head is zero. That's usually
okay because adding 0 is a no-op anyway. However, there's a corner case: if the tape head is already at
position 0, then adding to a cell to the left causes a write outside memory bounds. See tests/bug-1.b for
a testcase.

This change inserts a test for zero and a conditional jump before executing the instructions (unless the
cell value is known to be nonzero), so that the generated code will look like this, instead:

    movzbq (%rax), %rcx
    test   %cl, %cl
    jz     end
    add    %cl, -1(%rax)
    movb   $0x0, (%rax)
  end:

Hopefully, branch prediction works pretty well in this case, so that the conditional jump is not too
expensive.

This fixes #1.